Separate weblish from status dropdown
